all services routing with token validation
