+++ src/main/java/com/bizzlog/gateway/filters/AuthenticationPrefilter.java
@@ -1,120 +1,127 @@
-package com.bizzlog.gateway.filters;
-
-//import com.bizzlog.gateway.client.IAMClient;
-import com.bizzlog.gateway.client.UserResponse;
-import com.bizzlog.gateway.dto.ErrorResponseModel;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.bizzlog.gateway.utils.SecurityConstants;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.core.io.buffer.DataBufferFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.http.server.reactive.ServerHttpResponse;
-import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.WebClientResponseException;
-import org.springframework.web.server.ServerWebExchange;
-import reactor.core.publisher.Mono;
-
-import java.util.Date;
-import java.util.List;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
-
-@Order(Ordered.HIGHEST_PRECEDENCE)
-@Component
-@Slf4j
-public class AuthenticationPrefilter extends AbstractGatewayFilterFactory<AuthenticationPrefilter.Config> {
+//package com.bizzlog.gateway.filters;
+//
+////import com.bizzlog.gateway.client.IAMClient;
+//import com.bizzlog.gateway.client.UserResponse;
+//import com.bizzlog.gateway.dto.ErrorResponseModel;
+//import com.fasterxml.jackson.core.JsonProcessingException;
+//import com.fasterxml.jackson.databind.ObjectMapper;
+//import com.bizzlog.gateway.utils.SecurityConstants;
+//import lombok.NoArgsConstructor;
+//import lombok.extern.slf4j.Slf4j;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.beans.factory.annotation.Qualifier;
+//import org.springframework.cloud.gateway.filter.GatewayFilter;
+//import org.springframework.cloud.gateway.filter.GatewayFilterChain;
+//import org.springframework.cloud.gateway.filter.GlobalFilter;
+//import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
+//import org.springframework.core.Ordered;
+//import org.springframework.core.annotation.Order;
+//import org.springframework.core.io.buffer.DataBufferFactory;
+//import org.springframework.http.HttpStatus;
+//import org.springframework.http.server.reactive.ServerHttpRequest;
+//import org.springframework.http.server.reactive.ServerHttpResponse;
+//import org.springframework.stereotype.Component;
+//import org.springframework.web.reactive.function.client.WebClient;
+//import org.springframework.web.reactive.function.client.WebClientResponseException;
+//import org.springframework.web.server.ServerWebExchange;
+//import reactor.core.publisher.Mono;
+//
+//import java.util.Date;
+//import java.util.List;
+//import java.util.function.Predicate;
+//import java.util.stream.Collectors;
+//
+//@Order(Ordered.HIGHEST_PRECEDENCE)
+//@Component
+//@Slf4j
+//public class AuthenticationPrefilter extends AbstractGatewayFilterFactory<AuthenticationPrefilter.Config> {
+////
+////    @Autowired
+////    private IAMClient iamClient;
 //
 //    @Autowired
-//    private IAMClient iamClient;
-
-    @Autowired
-    @Qualifier("excludedUrls")
-    List<String> excludedUrls;
-    private final WebClient.Builder webClientBuilder;
-
-    public AuthenticationPrefilter(WebClient.Builder webClientBuilder) {
-        super(Config.class);
-        this.webClientBuilder=webClientBuilder;
-    }
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @Override
-    public GatewayFilter apply(Config config) {
-        return (exchange, chain) -> {
-            ServerHttpRequest request = exchange.getRequest();
-            log.info("**************************************************************************");
-            log.info("URL is - " + request.getURI().getPath());
-            String bearerToken = request.getHeaders().getFirst(SecurityConstants.HEADER);
-            log.info("Bearer Token: "+ bearerToken);
-
-            if(isSecured.test(request)) {
-                return webClientBuilder.build().get()
-                        .uri("http://localhost:8081/api/v1/validateToken")
-                        .header(SecurityConstants.HEADER, bearerToken)
-                        .retrieve().bodyToMono(UserResponse.class)
-                        .map(response -> {
-                            exchange.getRequest().mutate().header("username", response.getUserName());
-                            exchange.getRequest().mutate().header(SecurityConstants.ROLE, response.getRoles().stream().collect(Collectors.joining(",")));
-//                            exchange.getRequest().mutate().header("authorities", response.getAuthorities().stream().map(Authorities::getAuthority).reduce("", (a, b) -> a + "," + b));
-//                            exchange.getRequest().mutate().header("auth-token", response.getToken());
-
-                            return exchange;
-                        }).flatMap(chain::filter).onErrorResume(error -> {
-                            log.info("Error Happened");
-                            HttpStatus errorCode = null;
-                            String errorMsg = "";
-                            if (error instanceof WebClientResponseException) {
-                                WebClientResponseException webCLientException = (WebClientResponseException) error;
-                                errorCode =  HttpStatus.BAD_GATEWAY;//webCLientException.getStatusCode();
-                                errorMsg = webCLientException.getStatusText();
-
-                            } else {
-                                errorCode = HttpStatus.BAD_GATEWAY;
-                                errorMsg = HttpStatus.BAD_GATEWAY.getReasonPhrase();
-                            }
-//                            AuthorizationFilter.AUTH_FAILED_CODE
-                            return onError(exchange, String.valueOf(errorCode.value()) ,errorMsg, "JWT Authentication Failed", errorCode);
-                        });
-            }
-
-            return chain.filter(exchange);
-        };
-    }
-
-    public Predicate<ServerHttpRequest> isSecured = request -> excludedUrls.stream().noneMatch(uri -> request.getURI().getPath().contains(uri));
-    private Mono<Void> onError(ServerWebExchange exchange, String errCode, String err, String errDetails, HttpStatus httpStatus) {
-        DataBufferFactory dataBufferFactory = exchange.getResponse().bufferFactory();
-//        ObjectMapper objMapper = new ObjectMapper();
-        ServerHttpResponse response = exchange.getResponse();
-        response.setStatusCode(httpStatus);
-        try {
-            response.getHeaders().add("Content-Type", "application/json");
-            ErrorResponseModel data = new ErrorResponseModel(errCode, err, errDetails, null, new Date());
-            byte[] byteData = objectMapper.writeValueAsBytes(data);
-            return response.writeWith(Mono.just(byteData).map(t -> dataBufferFactory.wrap(t)));
-
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-
-        }
-        return response.setComplete();
-    }
-
-    @NoArgsConstructor
-    public static class Config {
-
-
-    }
-}
+//    @Qualifier("excludedUrls")
+//    List<String> excludedUrls;
+//    private final WebClient.Builder webClientBuilder;
+//
+//    public AuthenticationPrefilter(WebClient.Builder webClientBuilder) {
+//        super(Config.class);
+//        this.webClientBuilder=webClientBuilder;
+//    }
+//
+//    @Autowired
+//    private ObjectMapper objectMapper;
+//
+//    @Override
+//    public GatewayFilter apply(Config config) {
+//        log.info("**************************************************************************");
+//        log.info("inside apply method");
+//        log.info("**************************************************************************");
+//
+//        return (exchange, chain) -> {
+//            ServerHttpRequest request = exchange.getRequest();
+//            log.info("**************************************************************************");
+//            log.info("URL is - " + request.getURI().getPath());
+//            String bearerToken = request.getHeaders().getFirst(SecurityConstants.HEADER);
+//            log.info("Bearer Token: "+ bearerToken);
+//
+//            if(isSecured.test(request)) {
+//                return webClientBuilder.build().get()
+//                        .uri("http://localhost:8081/api/v1/validateToken")
+//                        .header(SecurityConstants.HEADER, bearerToken)
+//                        .retrieve().bodyToMono(UserResponse.class)
+//                        .map(response -> {
+//                            exchange.getRequest().mutate().header("username", response.getUserName());
+//                            exchange.getRequest().mutate().header(SecurityConstants.ROLE, response.getRoles().stream().collect(Collectors.joining(",")));
+////                            exchange.getRequest().mutate().header("authorities", response.getAuthorities().stream().map(Authorities::getAuthority).reduce("", (a, b) -> a + "," + b));
+////                            exchange.getRequest().mutate().header("auth-token", response.getToken());
+//
+//                            return exchange;
+//                        }).flatMap(chain::filter).onErrorResume(error -> {
+//                            log.info("Error Happened");
+//                            HttpStatus errorCode = null;
+//                            String errorMsg = "";
+//                            if (error instanceof WebClientResponseException) {
+//                                WebClientResponseException webCLientException = (WebClientResponseException) error;
+//                                errorCode =  HttpStatus.BAD_GATEWAY;//webCLientException.getStatusCode();
+//                                errorMsg = webCLientException.getStatusText();
+//
+//                            } else {
+//                                errorCode = HttpStatus.BAD_GATEWAY;
+//                                errorMsg = HttpStatus.BAD_GATEWAY.getReasonPhrase();
+//                            }
+////                            AuthorizationFilter.AUTH_FAILED_CODE
+//                            return onError(exchange, String.valueOf(errorCode.value()) ,errorMsg, "JWT Authentication Failed", errorCode);
+//                        });
+//            }
+//
+//            return chain.filter(exchange);
+//        };
+//    }
+//
+//    public Predicate<ServerHttpRequest> isSecured = request -> excludedUrls.stream().noneMatch(uri -> request.getURI().getPath().contains(uri));
+//    private Mono<Void> onError(ServerWebExchange exchange, String errCode, String err, String errDetails, HttpStatus httpStatus) {
+//        DataBufferFactory dataBufferFactory = exchange.getResponse().bufferFactory();
+////        ObjectMapper objMapper = new ObjectMapper();
+//        ServerHttpResponse response = exchange.getResponse();
+//        response.setStatusCode(httpStatus);
+//        try {
+//            response.getHeaders().add("Content-Type", "application/json");
+//            ErrorResponseModel data = new ErrorResponseModel(errCode, err, errDetails, null, new Date());
+//            byte[] byteData = objectMapper.writeValueAsBytes(data);
+//            return response.writeWith(Mono.just(byteData).map(t -> dataBufferFactory.wrap(t)));
+//
+//        } catch (JsonProcessingException e) {
+//            e.printStackTrace();
+//
+//        }
+//        return response.setComplete();
+//    }
+//
+//
+//    @NoArgsConstructor
+//    public static class Config {
+//
+//
+//    }
+//}
