+++ src/main/java/com/bizzlog/gateway/filters/AuthorizationFilter.java
@@ -13,6 +13,7 @@ import org.springframework.core.io.buffer.DataBufferFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.http.server.reactive.ServerHttpResponse;
+import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
 import reactor.core.publisher.Mono;
 
@@ -24,6 +25,7 @@ import java.util.function.Predicate;
 //@Component
 //@RefreshScope
 @Slf4j
+//@Component
 public class AuthorizationFilter implements GlobalFilter {
